isis: Address comments for BIER TLV
- Rename BierInfoTlv into BierInfoSubTlv and move to subtlvs/prefix/mod.rs
- Replace iterator with loop + fix comment
- Define minimum BierInfoSubTlv length in dedicated constant

See https://github.com/holo-routing/holo/pull/48#pullrequestreview-2645507200
