+++ holo-isis/src/packet/subtlvs/mod.rs
@@ -8,3 +8,4 @@
 //
 
 pub mod neighbor;
+pub mod prefix;
